--- src/constants.ts
@@ -1,4 +1,0 @@
-export const ID = 'id' as const;
-export const DEL_FLAG = 'delFlag' as const;
-export const STATUS_NORMAL = 0 as const;
-export const STATUS_REMOVED = 1 as const;
