+++ src/defs/constants.ts
@@ -1,0 +1,9 @@
+/**
+ * 实体类主键字段名
+ */
+export const ID = 'id' as const;
+
+/**
+ * 实体类逻辑删除标识字段名
+ */
+export const DEL_FLAG = 'delFlag' as const;
