+++ src/index.ts
@@ -1,6 +1,5 @@
-export * from './core';
-export * from './base';
-export * from './constants';
-export * from './error';
+export * from './dao';
+export * from './defs';
 export * from './types';
 export * from './utils';
+export * from './orm';
