+++ src/utils/beans.ts
@@ -1,6 +1,5 @@
 /**
  * 大小驼峰转下划线分割
- * @param name
  * @example camel2underline("createDate") === "create_date"
  * @example camel2underline("UpdateDate") === "update_date"
  * @returns
@@ -10,11 +9,11 @@ export const camel2underline = (name: string) => {
         return (index ? '_' : '') + sub.toLowerCase();
     });
 };
+
+/**
+ * 首字母转为大写
+ * @example upperFirstChar("name") === "Name"
+ */
 export const upperFirstChar = (str: string) => {
     return str.replace(/^./, (sub) => sub.toUpperCase());
-};
-
-export const useAutoIncrementId = () => {
-    let id = 0;
-    return () => ++id;
 };
